refactor(slf): extract FormulaHelper to its own TU (isolate exprtk)
Move the FormulaParams enum, kFormulaVars table, and the exprtk-backed
FormulaHelper out of ShadowCasterManager into FormulaHelper.{h,cpp}. The
header uses a void* pimpl so it stays exprtk-free; FormulaHelper.cpp is the
single TU that compiles the (huge) exprtk header. ShadowCasterManager.cpp is
now exprtk-free (the score/redraw formulas still call FormulaHelper::SetParam
etc. unchanged); the formula-editor UI keeps using kFormulaVars from the
header. Pure code move -- no behavior change. Plugin builds clean.

Also adds an opt-in cpp_tests_slow target (-DBUILD_CPP_SLOW_TESTS=ON, OFF by
default so the fast suite / PR gate is unaffected) with test_formulahelper.cpp
covering Parse/Calculate/Validate/Reparse, param binding, and error handling.

KNOWN ISSUE: cpp_tests_slow builds but currently registers 0 Catch2 test
cases under the local MSVC 14.50 toolchain (a trivial probe TEST_CASE in the
main TU also fails to register), so the tests don't yet execute -- a
target-level Catch2 registration problem to resolve, independent of the
extraction (which is plugin-validated) and the test code.
