+++ src/Features/LightLimitFix/ShadowCasterManager.cpp
@@ -15,140 +15,10 @@
 #include "../Upscaling.h"
 #include "../VR.h"
 
-#include <exprtk.hpp>
-
 namespace ShadowCasterManager
 {
-	// =========================================================================
-	// Formula evaluator (exprtk)
-	// =========================================================================
-
-	struct FormulaWrapper
-	{
-		exprtk::expression<double> expression;
-		exprtk::parser<double> parser;
-	};
-
-	static double s_formulaParams[kFormulaParam_Max];
-	static exprtk::symbol_table<double> s_symbolTable;
-	static bool s_formulaInited = false;
-
-	struct FormulaVarInfo
-	{
-		const char* name;
-		const char* description;
-		int32_t index;
-	};
-
-	// Single authoritative list of formula variables.
-	// Drives both symbol table registration and the formula editor help text.
-	static constexpr FormulaVarInfo kFormulaVars[] = {
-		{ "lightindex", "sequential index of this candidate light", kFormulaParam_LightIndex },
-		{ "lightintensity", "NiLight fade/intensity", kFormulaParam_LightIntensity },
-		{ "lightdistance", "camera-to-light distance (game units; 1 unit ~= 1.428 cm)", kFormulaParam_LightDistance },
-		{ "lightradius", "light radius/range (game units; 1 unit ~= 1.428 cm)", kFormulaParam_LightRadius },
-		{ "lightx", "light world X", kFormulaParam_LightX },
-		{ "lighty", "light world Y", kFormulaParam_LightY },
-		{ "lightz", "light world Z", kFormulaParam_LightZ },
-		{ "lightr", "diffuse red", kFormulaParam_LightR },
-		{ "lightg", "diffuse green", kFormulaParam_LightG },
-		{ "lightb", "diffuse blue", kFormulaParam_LightB },
-		{ "lightambientr", "ambient red", kFormulaParam_LightAmbientR },
-		{ "lightambientg", "ambient green", kFormulaParam_LightAmbientG },
-		{ "lightambientb", "ambient blue", kFormulaParam_LightAmbientB },
-		{ "lightchosenlastframe", "1 if this light held a slot last frame", kFormulaParam_LightChosenLastFrame },
-		{ "lightframessincerender", "frames since this light's slot was last actually rendered into the shadow atlas; 1e6 sentinel when never rendered or unassigned", kFormulaParam_LightFramesSinceRender },
-		{ "lightneverfades", "1 if lodFade disabled (permanent light)", kFormulaParam_LightNeverFades },
-		{ "lightportalstrict", "1 if portal-strict (always 1 for shadow casters)", kFormulaParam_LightPortalStrict },
-		{ "lightns", "1 if promoted from normal light (PromoteNormalToShadow)", kFormulaParam_LightNS },
-		{ "lightconverted", "1 if light is in the converted (non-shadow) slot range", kFormulaParam_LightConverted },
-		{ "lightdisplacement", "distance this light moved since its last shadow map render (game units; 0 when not yet tracked or in score formula)", kFormulaParam_LightDisplacement },
-		{ "playerlightdistance", "distance from the player character to the light (game units; falls back to lightdistance when player unavailable)", kFormulaParam_PlayerLightDistance },
-		{ "lightimportance", "contribution score: lum(diffuse*fade) * max(att_cam,att_plr) where att=(1-(dist/radius)^2)^2; 0 in score formula", kFormulaParam_LightImportance },
-		{ "lightisspot", "1 if this is a spot/frustum shadow light (BSShadowFrustumLight); 0 for omni / hemi / sun", kFormulaParam_LightIsSpot },
-		{ "lightspotvisible", "1 if the spot's cone plausibly reaches the camera frustum, 0 otherwise. Always 1 for non-spot lights so existing omni-only formulas are unaffected", kFormulaParam_LightSpotVisible },
-		{ "camerax", "camera world X", kFormulaParam_CameraX },
-		{ "cameray", "camera world Y", kFormulaParam_CameraY },
-		{ "cameraz", "camera world Z", kFormulaParam_CameraZ },
-		{ "isinterior", "1 in interior cells, 0 outdoors", kFormulaParam_IsInterior },
-		{ "timeofday", "in-game hour (0.0-24.0)", kFormulaParam_TimeOfDay },
-		{ "frametime", "EMA-smoothed frame time (ms)", kFormulaParam_FrameTime },
-		{ "frametarget", "90th-percentile recent frame time (ms) -- headroom ceiling", kFormulaParam_FrameTarget },
-		{ "stableframes", "consecutive frames EMA has been below frametarget", kFormulaParam_StableFrames },
-	};
-
-	static void InitFormulaSystem()
-	{
-		if (s_formulaInited)
-			return;
-		s_formulaInited = true;
-
-		memset(s_formulaParams, 0, sizeof(double) * kFormulaParam_Max);
-
-		for (const auto& v : kFormulaVars)
-			s_symbolTable.add_variable(v.name, s_formulaParams[v.index]);
-	}
-
-	FormulaHelper::FormulaHelper() :
-		_ptr(nullptr) { InitFormulaSystem(); }
-
-	FormulaHelper::~FormulaHelper()
-	{
-		if (_ptr)
-			delete static_cast<FormulaWrapper*>(_ptr);
-	}
-
-	bool FormulaHelper::Parse(const std::string& input)
-	{
-		if (_ptr)
-			return false;
-		auto* w = new FormulaWrapper();
-		w->expression.register_symbol_table(s_symbolTable);
-		// Defer the _ptr assignment until compile succeeds. Otherwise a
-		// failed compile leaves the helper in a "parsed" state (Calculate
-		// would evaluate an uncompiled expression and the early-return
-		// guard above would block subsequent Parse retries).
-		if (!w->parser.compile(input, w->expression)) {
-			delete w;
-			return false;
-		}
-		_ptr = w;
-		return true;
-	}
-
-	double FormulaHelper::Calculate()
-	{
-		auto* w = static_cast<FormulaWrapper*>(_ptr);
-		return w ? w->expression.value() : 0.0;
-	}
-
-	bool FormulaHelper::Reparse(const std::string& input)
-	{
-		std::string err;
-		if (!Validate(input, err))
-			return false;
-		if (_ptr)
-			delete static_cast<FormulaWrapper*>(_ptr);
-		_ptr = nullptr;
-		return Parse(input);
-	}
-
-	bool FormulaHelper::Validate(const std::string& input, std::string& errorOut)
-	{
-		InitFormulaSystem();
-		FormulaWrapper tmp;
-		tmp.expression.register_symbol_table(s_symbolTable);
-		if (tmp.parser.compile(input, tmp.expression))
-			return true;
-		if (tmp.parser.error_count() > 0)
-			errorOut = tmp.parser.get_error(0).diagnostic;
-		else
-			errorOut = "Unknown parse error";
-		return false;
-	}
-
-	void FormulaHelper::SetParam(int32_t index, double value) { s_formulaParams[index] = value; }
-	double FormulaHelper::GetParam(int32_t index) { return s_formulaParams[index]; }
+	// Formula evaluator (FormulaWrapper / kFormulaVars / FormulaHelper) moved to
+	// FormulaHelper.{h,cpp} to isolate the heavy exprtk header to one TU.
 
 	// =========================================================================
 	// Module-level state
